Make Tools controller prevent users from deleting arbitrary tools
Admins and tool creators are still allowed to delete tools, though.

+++ app/controllers/tools_controller.rb
@@ -49,16 +49,25 @@ class ToolsController < ApplicationController
   
   def destroy
     if tool = Tool.find(params[:id])
-      name = tool.name
-      if tool.destroy
-        flash[:notice] = "You've deleted the #{name}."
-        redirect_to tools_path
+      if current_user.can_administrate? or current_user == tool.user
+        name = tool.name
+        if tool.destroy
+          flash[:notice] = "You've deleted the #{name}."
+          redirect_to tools_path
+          return
+        else
+          flash[:alert] = "Whoops, something bad happened!"
+          redirect_to edit_tool_path(tool)
+          return
+        end  
       else
-        flash[:alert] = "Whoops, something bad happened!"
+        flash[:alert] = "You are not allowed to delete tools you don't own"
         redirect_to edit_tool_path(tool)
+        return
       end
     else
       flash[:alert] = "Could not find tool with id #{ params[:id] }."
+      return
     end
   end
